requirements.txt
# Driver Rating KZ Pro - Requirements

# Core
aiogram==3.15.0
asyncpg==0.30.0
python-dotenv==1.0.0

# Utilities
aiohttp==3.9.3

# Optional: Payment integrations
# stripe==7.10.0

# Optional: Analytics
# sentry-sdk==1.40.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
